src/api/requirements.txt
flake8==3.7.9
Flask==2.1.0
Flask-RESTful==0.3.9
flask-swagger-ui==3.25.0
Flask-Cors==3.0.8
tinydb==4.1.0
PyYAML==5.3.1
jinja2==3.0.3